Code V0.8.9

+++ 06-Code/PanesRumiñahui/src/ec/edu/espe/panesRumiñahui/model/Sale.java
@@ -1,22 +1,22 @@
 package ec.edu.espe.panesrumiñahui.model;
+
 /**
  *
  * @author Gabriel Baez, Techware, DCCO-ESPE
  */
 public class Sale {
-
-    private double totalSales;
+    private float totalSales;
     
 
-    public Sale(double totalSales) {
+    public Sale(float totalSales) {
         this.totalSales = totalSales;
     }
 
-    public double getTotalSales() {
+    public float getTotalSales() {
         return totalSales;
     }
 
-    public void setTotalSales(double totalSales) {
+    public void setTotalSales(float totalSales) {
         this.totalSales = totalSales;
     }
 
@@ -24,8 +24,5 @@ public class Sale {
     public String toString() {
         return "\n\tGanancias Diarias:" + "\n\t\tGanancia del Dia=" + getTotalSales();
     }
-    
 
-
-    
 }
